moved functions to request supported KNN properties from utils/knn_utils.h/.cpp to dimensionality_reduction/knn_support.h/.cpp
