+++ hdi/dimensionality_reduction/hierarchical_sne.cpp
@@ -41,18 +41,7 @@
 namespace hdi{
   namespace dr{
 
-	  int HierarchicalSNE_NrOfKnnAlgorithms()
-	  {
-          int numSupported = 1;
-#ifdef HNSWLIB_SUPPORTED
-          numSupported++;
-#endif
-          
-#ifdef __USE_ANNOY__
-          numSupported++;
-#endif
-		  return numSupported;
-	  }
+	  
 
     template class HierarchicalSNE<float,std::vector<std::map<uint32_t,float>>>;
    // template class HierarchicalSNE<double,std::vector<std::map<uint32_t,double>>>;
