+++ hdi/dimensionality_reduction/knn_support.h
@@ -1,6 +1,6 @@
 /*
  *
- * Copyright (c) 2020, BIOVAULT (Leiden University Medical Center, Delft University of Technology)
+ * Copyright (c) 2014, Nicola Pezzotti (Delft University of Technology)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -17,7 +17,7 @@
  *  its contributors may be used to endorse or promote products derived from
  *  this software without specific prior written permission.
  *
- * THIS SOFTWARE IS PROVIDED BY BIOVAULT ''AS IS'' AND ANY EXPRESS
+ * THIS SOFTWARE IS PROVIDED BY NICOLA PEZZOTTI ''AS IS'' AND ANY EXPRESS
  * OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
  * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
  * EVENT SHALL NICOLA PEZZOTTI BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
@@ -30,50 +30,20 @@
  *
  */
 
-#include "knn_utils.h"
+#ifndef KNN_SUPPORT_H
+#define KNN_SUPPORT_H
+#include <vector>
 
+namespace hdi {
+  namespace dr {
+    int HierarchicalSNE_NrOfKnnAlgorithms();
 
+    //! returns which knn libraries are supported.
+    std::vector<int> supported_knn_libraries();
 
-namespace hdi{
-  namespace utils{
-           
-    std::vector<int> supported_knn_libraries()
-    {
-      std::vector<int> result(1, KNN_FLANN);
-#ifdef HNSWLIB_SUPPORTED
-      result.push_back(KNN_HNSW);
-#endif
-#ifdef __USE_ANNOY__
-      result.push_back(KNN_ANNOY);
-#endif
-      return result;
-    }
-
-  
-    std::vector<int> supported_knn_library_distance_metrics(int knn_lib)
-    {
-      switch (knn_lib)
-      {
-        case KNN_FLANN: return std::vector<int>(1, KNN_METRIC_EUCLIDEAN);
-        case KNN_HNSW: {
-          std::vector<int> result(2);
-          result[0] = KNN_METRIC_EUCLIDEAN;
-          result[1] = KNN_METRIC_INNER_PRODUCT;
-          return result;
-
-        }
-        case KNN_ANNOY: {
-          std::vector<int> result(5);
-          result[0] = KNN_METRIC_EUCLIDEAN;
-          result[1] = KNN_METRIC_COSINE;
-          result[2] = KNN_METRIC_MANHATTAN;
-          result[3] = KNN_METRIC_DOT;
-          return result;
-        }
-
-        default: throw std::out_of_range("knn_lib value out of range");
-      }
-    }
+    //! returns a vector with distance metrics supported for a specific knn library
+    std::vector<int> supported_knn_library_distance_metrics(int knn_lib);
   }
 }
 
+#endif
